dealer-rick/requirements.txt
discord.py
python-dotenv
aiosqlite
Pillow